refactor - LoginCover elements
