+++ src/components/LoginCover/elements/CoverLogo.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+const CoverLogo = styled.h1`
+    font-family: Passion One;
+    font-size: 106px;
+    font-weight: 700;
+    letter-spacing: 4px;
+
+    @media (max-width: 937px) {
+        font-size: 76px;
+        margin-bottom: -8px;
+    }
+`;
+
+export default CoverLogo;
