+++ src/pages/sign-up/SignUp.js
@@ -1,6 +1,6 @@
 import SignForm from "../../components/SignForm/SignForm";
 
-import LoginCover from "../../components/LoginCover";
+import LoginCover from "../../components/LoginCover/LoginCover";
 
 import styled from "styled-components";
